Add documentation; Add functions allowing for more control over functionality of AutoSwitch & Consumable; Add CI workflow
