+++ lib/src/main/kotlin/com/w2sv/kotlinutils/TypeAliases.kt
