+++ lib/src/main/kotlin/com/w2sv/kotlinutils/extensions/Boolean.kt
@@ -5,5 +5,5 @@ package com.w2sv.kotlinutils.extensions
 fun Boolean.toInt(): Int =
     if (equals(true)) 1 else 0
 
-fun Boolean.toNonZeroInt(): Int =
+fun Boolean.toNonZeroInt(): Int =  // TODO: come up with better name
     if (equals(true)) 1 else -1
