+++ lib/src/main/kotlin/com/w2sv/kotlinutils/extensions/Enum.kt
@@ -2,5 +2,5 @@
 
 package com.w2sv.kotlinutils.extensions
 
-val Enum<*>.nonZeroOrdinal: Int get() =
+val Enum<*>.nonZeroOrdinal: Int get() =  // TODO: come up with better name
     ordinal + 1
