+++ lib/src/main/kotlin/com/w2sv/kotlinutils/extensions/Float.kt
