+++ lib/src/test/kotlin/com/w2sv/kotlinutils/delegates/ConsumableTest.kt
