+++ lib/src/test/kotlin/com/w2sv/kotlinutils/delegates/MapObservableKtTest.kt
@@ -4,14 +4,14 @@ import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 
-internal class MapObserverKtTest {
+internal class MapObservableKtTest {
 
     companion object {
         var nObserverCalls = 0
     }
 
     private val map = mutableMapOf("property" to 69)
-    private var property by MapObserver(map) { _, _, _ ->
+    private var property by MapObservable(map) { _, _, _ ->
         nObserverCalls += 1
     }
