+++ lib/src/test/kotlin/com/w2sv/kotlinutils/extensions/FloatKtTest.kt
